# backend/docs/swagger.yaml
definitions:
  dto.AssessmentHistoryEntryDTO:
    properties:
      bmi:
        type: number
      bmr:
        type: number
      body_fat:
        type: number
      id:
        type: string
      recorded_at:
        type: string
      tdee:
        type: number
      weight:
        type: number
    type: object
...
  dto.CreateAssessmentInputDTO:
    properties:
      activity_level:
        type: string
      body_fat:
        type: number
      height:
        type: number
      user_id:
        type: string
      weight:
        type: number
    type: object
  dto.CreateDietPlanInputDTO:
    properties:
      calorie_intensity:
        type: string
      fat_intensity:
        type: string
      protein_intensity:
        type: string
      protocol_id:
        type: string
    type: object
...
  dto.DietPlanAdherenceOutputDTO:
    properties:
      calories:
        $ref: '#/definitions/dto.MacroAdherenceDTO'
      carbs:
        $ref: '#/definitions/dto.MacroAdherenceDTO'
      date:
        type: string
      diet_plan_id:
        type: string
      fat:
        $ref: '#/definitions/dto.MacroAdherenceDTO'
      global_adherence:
        description: average of individual ratios; > 1.0 means over-target overall
        type: number
      protein:
        $ref: '#/definitions/dto.MacroAdherenceDTO'
    type: object
  dto.DietPlanOutputDTO:
    properties:
      calorie_intensity:
        type: string
      calories:
        type: number
      carbs:
        type: number
      fat:
        type: number
      fat_intensity:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      protein:
        type: number
      protein_intensity:
        type: string
      protocol_id:
        type: string
      water:
        type: number
    type: object
  dto.MacroAdherenceDTO:
    properties:
      adherence:
        description: ratio consumed/target; > 1.0 means over-target
        type: number
      consumed:
        type: number
      target:
        type: number
    type: object
...
  dto.UpdateAssessmentInputDTO:
    properties:
      activity_level:
        type: string
      body_fat:
        type: number
      height:
        type: number
      weight:
        type: number
    type: object
  dto.UpdateDietPlanInputDTO:
    properties:
      calorie_intensity:
        type: string
      calories:
        type: number
      carbs:
        type: number
      fat:
        type: number
      fat_intensity:
        type: string
      is_active:
        type: boolean
      protein:
        type: number
      protein_intensity:
        type: string
      water:
        type: number
    type: object
...
  entity.Assessment:
    properties:
      activityLevel:
        type: string
      bmi:
        format: float64
        type: number
      bmr:
        format: float64
        type: number
      bodyFat:
        format: float64
        type: number
      height:
        format: float64
        type: number
      id:
        type: string
      recordedAt:
        type: string
      tdee:
        format: float64
        type: number
      userID:
        type: string
      weight:
        format: float64
        type: number
    type: object
...
info:
  contact: {}
paths:
  /assessment:
    get:
      description: Retrieves a list of all registered assessment profiles sorted by
        name.
      parameters:
      - description: User UUID to filter assessments by user
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Assessment'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all assessments
      tags:
      - assessment
    post:
      consumes:
      - application/json
      description: Creates an assessment
      parameters:
      - description: Assessment Data
        in: body
        name: assessment
        required: true
        schema:
          $ref: '#/definitions/dto.CreateAssessmentInputDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Assessment created successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input data
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new assessment
      tags:
      - assessment
  /assessment/{id}:
    delete:
      description: Permanently removes an assessment and their associated data from
        the system.
      parameters:
      - description: Assessment UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete assessment profile
      tags:
      - assessment
    get:
      description: Retrieves the basic information of a specific assessment.
      parameters:
      - description: Assessment UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Assessment'
        "404":
          description: Assessment not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get assessment by ID
      tags:
      - assessment
    put:
      consumes:
      - application/json
      description: Updates the weight, height, or other data of an existing assessment.
      parameters:
      - description: Assessment UUID
        in: path
        name: id
        required: true
        type: string
      - description: Updated Assessment Data
        in: body
        name: assessment
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateAssessmentInputDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Assessment updated successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid update data
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update assessment profile
      tags:
      - assessment
  /assessment/history:
    get:
      parameters:
      - description: User UUID
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.AssessmentHistoryEntryDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Assessment history (shape evolution timeline)
      tags:
      - assessment
  /diet-plan:
    get:
      parameters:
      - description: Protocol UUID
        in: query
        name: protocol_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.DietPlanOutputDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List diet plans by protocol
      tags:
      - diet_plan
    post:
      consumes:
      - application/json
      parameters:
      - description: Diet Plan Data
        in: body
        name: diet_plan
        required: true
        schema:
          $ref: '#/definitions/dto.CreateDietPlanInputDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new diet plan
      tags:
      - diet_plan
  /diet-plan/{id}:
    delete:
      parameters:
      - description: Diet Plan UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete diet plan
      tags:
      - diet_plan
    get:
      parameters:
      - description: Diet Plan UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DietPlanOutputDTO'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get diet plan by ID
      tags:
      - diet_plan
    put:
      consumes:
      - application/json
      parameters:
      - description: Diet Plan UUID
        in: path
        name: id
        required: true
        type: string
      - description: Updated Diet Plan Data
        in: body
        name: diet_plan
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateDietPlanInputDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update diet plan
      tags:
      - diet_plan
  /diet-plan/adherence:
    get:
      parameters:
      - description: Diet Plan UUID
        in: query
        name: diet_plan_id
        required: true
        type: string
      - description: User UUID
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DietPlanAdherenceOutputDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Diet plan adherence for today
      tags:
      - diet_plan
...
swagger: "2.0"