# backend/internal/infra/database/postgresql/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package sqlc

import (
	"database/sql"
	"encoding/json"
	"time"

	"github.com/google/uuid"
)

type Assessment struct {
	ID            uuid.UUID    `json:"id"`
	UserID        uuid.UUID    `json:"user_id"`
	Weight        float64      `json:"weight"`
	Height        float64      `json:"height"`
	BodyFat       float64      `json:"body_fat"`
	ActivityLevel string       `json:"activity_level"`
	Bmi           float64      `json:"bmi"`
	Bmr           float64      `json:"bmr"`
	Tdee          float64      `json:"tdee"`
	CreatedAt     sql.NullTime `json:"created_at"`
	UpdatedAt     sql.NullTime `json:"updated_at"`
	DeletedAt     sql.NullTime `json:"deleted_at"`
}

type DailyMacroSummary struct {
	UserID        uuid.UUID `json:"user_id"`
	LogDate       time.Time `json:"log_date"`
	TotalProtein  string    `json:"total_protein"`
	TotalCarbs    string    `json:"total_carbs"`
	TotalFat      string    `json:"total_fat"`
	TotalCalories string    `json:"total_calories"`
}

type DietPlan struct {
	ID               uuid.UUID    `json:"id"`
	ProtocolID       uuid.UUID    `json:"protocol_id"`
	CalorieIntensity string       `json:"calorie_intensity"`
	ProteinIntensity string       `json:"protein_intensity"`
	FatIntensity     string       `json:"fat_intensity"`
	Protein          float64      `json:"protein"`
	Carbs            float64      `json:"carbs"`
	Fat              float64      `json:"fat"`
	Calories         float64      `json:"calories"`
	Water            float64      `json:"water"`
	IsActive         sql.NullBool `json:"is_active"`
	CreatedAt        sql.NullTime `json:"created_at"`
	UpdatedAt        sql.NullTime `json:"updated_at"`
	DeletedAt        sql.NullTime `json:"deleted_at"`
}

type Goal struct {
	ID           uuid.UUID    `json:"id"`
	ProjectID    uuid.UUID    `json:"project_id"`
	Name         string       `json:"name"`
	StrategyType string       `json:"strategy_type"`
	IsActive     sql.NullBool `json:"is_active"`
	CreatedAt    sql.NullTime `json:"created_at"`
	UpdatedAt    sql.NullTime `json:"updated_at"`
	DeletedAt    sql.NullTime `json:"deleted_at"`
}

type MacroEstimation struct {
	ID              uuid.UUID       `json:"id"`
	UserID          uuid.UUID       `json:"user_id"`
	MealDescription string          `json:"meal_description"`
	MealItems       json.RawMessage `json:"meal_items"`
	CreatedAt       sql.NullTime    `json:"created_at"`
	UpdatedAt       sql.NullTime    `json:"updated_at"`
	DeletedAt       sql.NullTime    `json:"deleted_at"`
}

type MealLog struct {
	ID              uuid.UUID       `json:"id"`
	UserID          uuid.UUID       `json:"user_id"`
	MealTitle       string          `json:"meal_title"`
	MealDescription string          `json:"meal_description"`
	MealItems       json.RawMessage `json:"meal_items"`
	CreatedAt       sql.NullTime    `json:"created_at"`
	UpdatedAt       sql.NullTime    `json:"updated_at"`
	DeletedAt       sql.NullTime    `json:"deleted_at"`
}

type Project struct {
	ID        uuid.UUID    `json:"id"`
	UserID    uuid.UUID    `json:"user_id"`
	Name      string       `json:"name"`
	IsActive  sql.NullBool `json:"is_active"`
	CreatedAt sql.NullTime `json:"created_at"`
	UpdatedAt sql.NullTime `json:"updated_at"`
	DeletedAt sql.NullTime `json:"deleted_at"`
}

type Protocol struct {
	ID        uuid.UUID    `json:"id"`
	GoalID    uuid.UUID    `json:"goal_id"`
	Name      string       `json:"name"`
	IsActive  sql.NullBool `json:"is_active"`
	CreatedAt sql.NullTime `json:"created_at"`
	UpdatedAt sql.NullTime `json:"updated_at"`
	DeletedAt sql.NullTime `json:"deleted_at"`
}

type User struct {
	ID            uuid.UUID    `json:"id"`
	Name          string       `json:"name"`
	Email         string       `json:"email"`
	BiologicalSex string       `json:"biological_sex"`
	BirthDate     time.Time    `json:"birth_date"`
	CreatedAt     sql.NullTime `json:"created_at"`
	UpdatedAt     sql.NullTime `json:"updated_at"`
	DeletedAt     sql.NullTime `json:"deleted_at"`
}

type UserDailyPerformance struct {
	UserID               uuid.UUID `json:"user_id"`
	UserName             string    `json:"user_name"`
	LogDate              time.Time `json:"log_date"`
	ConsumedProtein      string    `json:"consumed_protein"`
	Protein              float64   `json:"protein"`
	ConsumedCalories     string    `json:"consumed_calories"`
	Calories             float64   `json:"calories"`
	Water                float64   `json:"water"`
	CaloriesAdherencePct int32     `json:"calories_adherence_pct"`
}